Create separate storage for used ports
