+++ src/kernels/common/usedPorts.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+
+// Keeps track of all ports used by Kernels and other processes spawned by Kernels and related code
+export const UsedPorts = new Set<number>();
