+++ src/kernels/portAttributeProvider.node.ts
@@ -4,11 +4,10 @@
 import { inject, injectable } from 'inversify';
 import { workspace } from 'vscode';
 import { CancellationToken, PortAttributes, PortAttributesProvider, PortAutoForwardAction } from 'vscode';
-import { NotebookStarter } from '../../jupyter/launcher/notebookStarter.node';
-import { KernelLauncher } from '../launcher/kernelLauncher.node';
-import { IExtensionSyncActivationService } from '../../../platform/activation/types';
-import { traceError } from '../../../platform/logging';
-import { IDisposableRegistry } from '../../../platform/common/types';
+import { IExtensionSyncActivationService } from '../platform/activation/types';
+import { traceError } from '../platform/logging';
+import { IDisposableRegistry } from '../platform/common/types';
+import { UsedPorts } from './common/usedPorts';
 
 /**
  * Used to determine how ports can be used when creating a raw kernel.
@@ -31,7 +30,7 @@ export class PortAttributesProviders implements PortAttributesProvider, IExtensi
         _token: CancellationToken
     ): PortAttributes | undefined {
         try {
-            if (KernelLauncher.usedPorts.includes(port) || NotebookStarter.usedPorts.includes(port)) {
+            if (UsedPorts.has(port)) {
                 return new PortAttributes(port, PortAutoForwardAction.Ignore);
             }
         } catch (ex) {
